produk admin

+++ app/actions/pageadmin/produk/galeri_delete.php
@@ -2,7 +2,7 @@
 $galeri=  _select_unique_result("select * from galeri_produk where idGaleri=$_GET[idGaleri]");
 $is_success= _query("delete from galeri_produk where idGaleri=$_GET[idGaleri]");
 if($is_success){
-    if($galeri['format']=='video'){
+    if($galeri['format']=='video'  && file_exists($g['file'])){
         unlink($galeri['file']);
     }
     $_SESSION['success']="Data galeri berhasil dihapus";
